Feature/serializor

+++ flipflOS/Kernel.cs
@@ -4,17 +4,22 @@ using System.Linq;
 using System.IO;
 using System.Text;
 using Sys = Cosmos.System;
+using Cosmos.System.FileSystem.VFS;
 
 namespace flipflOS
 {
     public class Kernel : Sys.Kernel
     {
+        public static Sys.FileSystem.CosmosVFS fs1;
         DateTime start;
         Memory mem = new Memory(); //initialisieren aller Variablen ofc
         public Directory currentdir;
 
         protected override void BeforeRun()
         {
+            fs1 = new Sys.FileSystem.CosmosVFS();
+            Sys.FileSystem.VFS.VFSManager.RegisterVFS(fs1);
+
             createRoot(); //erstellt für DateiSystem das Root verzeichnis, sowie weitere
             Console.Clear();
             loadingScreen(5);
@@ -150,6 +155,9 @@ namespace flipflOS
                     case "mkdir":
                         makeDirectory(args);
                         break;
+                    case "removeDir":
+                        removeDirectory(args);
+                        break;
                     case "touch":
                         makeFile(args);
                         break;
@@ -218,6 +226,21 @@ namespace flipflOS
                     return;
                 }
             currentdir.createSubDirectory(args[1]);
+
+            Serializer.createDirectory(currentdir,args[1]);
+        }
+        public void removeDirectory(String[] args)
+        {
+            if (args.Length <= 1)
+            {
+                Console.WriteLine("Not enough Arguments. Syntax: mkdir dir");
+                return;
+            }
+            if (Serializer.deleteDirectory(currentdir, args[1]))
+            {
+                currentdir.removeSubDirectory(args[1]);
+            }
+
         }
         public void makeFile(String[] args)
         {
@@ -257,7 +280,7 @@ namespace flipflOS
             {
                 changeDir(seperatedbyslash[i]); //geht zum Pfad wo die Datei ist
             }
-            currentdir.deleteFile(file); //löscht File
+            if(Serializer.deleteFile(currentdir, file)) currentdir.deleteFile(file); //löscht File
             currentdir = startingdir; //geht wieder zum Startverzeichnis
         }
         public void moveFile(String[] args)
@@ -307,10 +330,7 @@ namespace flipflOS
         }
         public void createRoot()
         { //erstellt alle nötigen verzeichnisse für das dateisystem
-            currentdir = new Directory(null, null, "root");
-            currentdir.createSubDirectory("home");
-            currentdir.createSubDirectory("users");
-            currentdir.createSubDirectory("var");
+            currentdir = Serializer.createRoot();
         }
         public static void ClearCurrentLine()
         { //stellt sicher, dass während der navigation mit pfeiltasten keine neue Zeile begonnen wird.
@@ -386,9 +406,9 @@ namespace flipflOS
             Sleep(2000);
             Directory.File file = currentdir.getFile(fileString);
             Directory.File newfile = new FileEditor().startFileeditor(file);
-            currentdir.deleteFile(file.name);
-            currentdir.addFile(newfile);
-
+            if(Serializer.deleteFile(currentdir,file.name)) currentdir.deleteFile(file.name);
+            if(Serializer.saveFile(currentdir,file.name,newfile.content)) currentdir.addFile(newfile);
+            
             currentdir = startingdir; // zurück zum ersten Directory
         }
         public void read(String[] args)
